fix: 连接密码包含@字符时连接异常 add: 存储桶上传文件功能 add: 存储桶下载文件功能

+++ src/main/java/cn/oyzh/easymongo/mongo/MongoDatabase.java
@@ -9,7 +9,7 @@ public class MongoDatabase {
 
     private String name;
 
-    private Long sizeOnDisk;
+    private Double sizeOnDisk;
 
     public String getName() {
         return name;
@@ -19,11 +19,11 @@ public class MongoDatabase {
         this.name = name;
     }
 
-    public Long getSizeOnDisk() {
+    public Double getSizeOnDisk() {
         return sizeOnDisk;
     }
 
-    public void setSizeOnDisk(Long sizeOnDisk) {
+    public void setSizeOnDisk(Double sizeOnDisk) {
         this.sizeOnDisk = sizeOnDisk;
     }
 }
